[exporter/logicmonitorexporter] Support log resource mapping config (#29733)
**Description:** 
This PR adds support for resource mapping config for logs. 
To make the resource mapping for logs flexible, LogicMonitor platform
expects some params to be present in the log ingestion request. We have
changed the exporter configuration to accept that configuration.

**Link to tracking Issue:** #29732

**Testing:** Unit test cases were added. Also did the functional testing
for the log ingestion with the newly added config.

**Documentation:** 

As per the LogicMonitor's [Log Ingestion
documentation](https://www.logicmonitor.com/support/lm-logs/sending-logs-to-the-lm-logs-ingestion-api),
if more than one resource property exists, only the first property will
be mapped. In case of OTLP logs, there can be multiple resource
attributes and its order also cannot be guaranteed.

Recently, the LogicMonitor has made the resource mapping for logs more
flexible. With that, any of the resource attributes present in the log
matches can be considered for resource mapping. But, this is not the
default behaviour. In order to make the resource mapping flexible, you
can configure the resource_mapping_op in the LogicMonitor's exporter.

 ```
 exporters:
    logicmonitor:
     ...
      logs:
        resource_mapping_op: "OR"
```
The value for resource_mapping_op can be AND or OR. The values are case-insensitive.