[cmd/telemetrygen] Add TLS/mTLS support to telemetrygen (#29681)
**Description:** 
Implementing robust security measures is essential for any tracing
ingestion service and infrastructure. This adds TLS/mTLS support to
telemetrygen traces so that it can be used to test the security
implementation. To illustrate the usage, a new example, `secure-tracing`
is added to examples collection.

**Link to tracking Issue:** <Issue number if applicable>

**Testing:** 
Tested with the `secure-tracing` example.

**Documentation:** 
- telemetrygen README has been updated with a link to secure-tracing. 
- secure-tracing README contains setup of a testing environment and test
script via telemetrygen